# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "http://80.78.246.59/Stech/public/api/getCategory": {
            "get": {
                "tags": [
                    "GetParametrForProject"
                ],
                "summary": "Get list of Category",
                "description": "Returns list of Category",
                "operationId": "getCategory",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "http://80.78.246.59/Stech/public/api/createMessage": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "createMessage",
                "description": "createMessage",
                "operationId": "createMessage",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "type": "string",
                                        "format": "receiver_id",
                                        "example": "1"
                                    },
                                    "message": {
                                        "type": "string",
                                        "format": "message",
                                        "example": "Грузо-перевозки"
                                    },
                                    "file": {
                                        "type": "file",
                                        "format": "file",
                                        "example": "photo.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}